+++ spec/osm/api_spec.rb
@@ -878,6 +878,44 @@ describe "API" do
       })).should be_false
     end
 
+    it "Delete an event (succeded)" do
+      url = 'https://www.onlinescoutmanager.co.uk/events.php?action=deleteEvent&sectionid=1&eventid=2'
+      post_data = {
+        'apiid' => @api_config[:api_id],
+        'token' => @api_config[:api_token],
+        'userid' => 'user',
+        'secret' => 'secret',
+      }
+
+      api = Osm::Api.new('user', 'secret')
+      HTTParty.should_receive(:post).with(url, {:body => post_data}) { DummyHttpResult.new(:response=>{:code=>'200', :body=>'{"ok":true}'}) }
+
+      api.delete_event(Osm::Event.new({
+        :section_id => 1,
+        :name => 'Test event',
+        :id => 2
+      })).should be_true
+    end
+
+    it "Delete an event (failed)" do
+      url = 'https://www.onlinescoutmanager.co.uk/events.php?action=deleteEvent&sectionid=1&eventid=2'
+      post_data = {
+        'apiid' => @api_config[:api_id],
+        'token' => @api_config[:api_token],
+        'userid' => 'user',
+        'secret' => 'secret',
+      }
+
+      api = Osm::Api.new('user', 'secret')
+      HTTParty.should_receive(:post).with(url, {:body => post_data}) { DummyHttpResult.new(:response=>{:code=>'200', :body=>'{"ok":false}'}) }
+
+      api.delete_event(Osm::Event.new({
+        :section_id => 1,
+        :name => 'Test event',
+        :id => 2
+      })).should be_false
+    end
+
 
     it "Create a term (succeded)" do
       url = 'https://www.onlinescoutmanager.co.uk/users.php?action=addTerm&sectionid=1'