+++ CHANGELOG.md
@@ -3,6 +3,7 @@
   * Event:
     * Add api.ceate\_event
     * Add api.update\_event
+    * Add api.delete\_event
   * Event attendance:
     * Add api.get\_event
     * Add api.get\_event\_fields